Updated DAO

+++ src/main/java/com/imdbsystem/dao/RatingRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.imdbsystem.model.Rating;
 
-public interface RatingRepository extends CrudRepository<Rating, Integer> {
+public interface RatingRepository extends CrudRepository<Rating, Long> {
 
 }
