+++ src/main/java/com/imdbsystem/dao/ShowRespository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.imdbsystem.model.Show;
 
-public interface ShowRespository extends CrudRepository<Show, Integer> {
+public interface ShowRespository extends CrudRepository<Show, Long> {
 
 }
